Create component to list joined missions and rockets
